+++ iotdb-core/node-commons/src/test/java/org/apache/iotdb/commons/pipe/datastructure/LinkedListMessageQueueTest.java
@@ -17,9 +17,9 @@
  * under the License.
  */
 
-package org.apache.iotdb.commons.pipe;
+package org.apache.iotdb.commons.pipe.datastructure;
 
-import org.apache.iotdb.commons.pipe.schema.LinkedListMessageQueue;
+import org.apache.iotdb.commons.pipe.datastructure.LinkedListMessageQueue;
 
 import org.awaitility.Awaitility;
 import org.junit.Assert;
